# Conflicts:
#	compose.yaml

+++ src/test/java/white/monster/energy/adventurebackend/bookingTest/BookingRepositoryTest.java
@@ -14,15 +14,22 @@ import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+/**
+ * Test class for BookingRepository.
+ * Uses @DataJpaTest to configure in-memory database and JPA components for testing.
+ */
 @DataJpaTest
 public class BookingRepositoryTest {
 
+    // The BookingRepository to be tested
     @Autowired
     private BookingRepository bookingRepository;
 
+    // EntityManager for setting up test data
     @Autowired
     private EntityManager entityManager;
 
+    /** Set up test data before each test */
     @BeforeEach
     void setUp() throws Exception {
         // Arrange: Persist test bookings
@@ -32,6 +39,7 @@ public class BookingRepositoryTest {
         entityManager.flush();
     }
 
+    // Helper method to create a Booking instance with specified attributes
     private Booking createBooking(String type, String status, int visitorId, LocalDateTime start, LocalDateTime end, BigDecimal price) throws Exception {
         Booking b = new Booking();
         var clazz = b.getClass();
@@ -62,38 +70,50 @@ public class BookingRepositoryTest {
 
         var totalPriceField = clazz.getDeclaredField("totalPrice");
         totalPriceField.setAccessible(true);
-        totalPriceField.set(b, price);
+        totalPriceField.set(b, price.doubleValue());
 
         return b;
     }
 
-
+    /**
+     * Test method for BookingRepository.findByVisitorId.
+     * Verifies that bookings can be retrieved by visitor ID.
+     */
     @Test
     void testFindByVisitorId() {
-        // Act
+        // Act: Retrieve bookings for visitor ID 1
         List<Booking> bookings = bookingRepository.findByVisitorId(1);
 
-        // Assert
+        // Assert: Verify that 2 bookings are found
         assertEquals(2, bookings.size());
     }
 
+    /**
+     * Test method for BookingRepository.findByStatusIn.
+     * Verifies that bookings can be retrieved by a list of statuses.
+     */
     @Test
     void testFindByStatusIn() {
-        // Act
+        // Act: Retrieve bookings with status "HOLD" or "CONFIRMED"
         List<Booking> bookings = bookingRepository.findByStatusIn(List.of("HOLD", "CONFIRMED"), org.springframework.data.domain.PageRequest.of(0, 10)).getContent();
 
-        // Assert
+        // Assert: Verify that 2 bookings are found
         assertEquals(2, bookings.size());
     }
 
+    /**
+     * Test method for BookingRepository.findByStartTimeBetween.
+     * Verifies that bookings can be retrieved within a specific start time range.
+     */
     @Test
     void testFindByStartTimeBetween() {
-        // Act
+        // Act: Define time range and retrieve bookings
         LocalDateTime from = LocalDateTime.now().minusDays(2);
         LocalDateTime to = LocalDateTime.now().plusDays(2);
+        // List bookings that start between 'from' and 'to' in order to include the first two bookings
         List<Booking> bookings = bookingRepository.findByStartTimeBetween(from, to, org.springframework.data.domain.PageRequest.of(0, 10)).getContent();
 
-        // Assert
+        // Assert: Verify that 2 bookings are found
         assertEquals(2, bookings.size());
     }
 }
